Fix node-sass calling syntax to correct output
Also expand the output so it's more readable. I'm doing this so I can
use a separate tool for handling minification at a later step. I'm not
sure that this is the right way to go about handling minification, but
separation of concerns seems like a good thing, and I want to avoid
using my butter knife as a screwdriver just cause in order to save a
trip to the kitchen.