fix: :bug: Clear red alert when no longer in danger
Now I clear red alert when the stage is cleared.
I also now differentiate between ghosts and monsters when
checking the length of a queue.
I also always check to see if all fields are under control whenever
a monster is cleared instead of only when queue was out of control.